+++ Runtime/ContinuousDistributions/BatesDistribution/GeneratorProviders/BatesGeneratorDependentSimpleProvider.cs
@@ -115,6 +115,13 @@ namespace Zor.RandomGenerators.ContinuousDistributions
 			}
 		}
 
+		/// <inheritdoc/>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public override void DropSharedGenerator()
+		{
+			m_sharedGenerator = null;
+		}
+
 		private void OnValidate()
 		{
 			m_sharedGenerator = null;