+++ Runtime/ContinuousDistributions/DistributionModificators/Clamp/ModificatorProviders/ClampModificatorProvider.cs
@@ -126,6 +126,13 @@ namespace Zor.RandomGenerators.ContinuousDistributions.DistributionModificators
 			}
 		}
 
+		/// <inheritdoc/>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public override void DropSharedGenerator()
+		{
+			m_sharedModificator = null;
+		}
+
 		private void OnValidate()
 		{
 			m_sharedModificator = null;