+++ Runtime/ContinuousDistributions/SharpDistribution/GeneratorProviders/SharpGeneratorProvider.cs
@@ -92,6 +92,13 @@ namespace Zor.RandomGenerators.ContinuousDistributions
 			}
 		}
 
+		/// <inheritdoc/>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public override void DropSharedGenerator()
+		{
+			m_sharedGenerator = null;
+		}
+
 		private void OnValidate()
 		{
 			m_sharedGenerator = null;