+++ Runtime/DiscreteDistributions/BinomialDistribution/GeneratorProviders/BinomialGeneratorDependentProvider.cs
@@ -149,6 +149,13 @@ namespace Zor.RandomGenerators.DiscreteDistributions
 			}
 		}
 
+		/// <inheritdoc/>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public override void DropSharedGenerator()
+		{
+			m_sharedGenerator = null;
+		}
+
 		private void OnValidate()
 		{
 			m_sharedGenerator = null;