+++ Runtime/DiscreteDistributions/DistributionFilters/DescendantSequenceFilters/FilterProviders/DescendantSequenceFilterProvider.cs
@@ -91,6 +91,13 @@ namespace Zor.RandomGenerators.DiscreteDistributions.DistributionFilters
 			}
 		}
 
+		/// <inheritdoc/>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public sealed override void DropSharedFilter()
+		{
+			m_sharedFilter = null;
+		}
+
 		private void OnValidate()
 		{
 			m_sharedFilter = null;