+++ Runtime/DiscreteDistributions/PoissonDistribution/GeneratorProviders/PoissonGeneratorSimpleProvider.cs
@@ -90,6 +90,13 @@ namespace Zor.RandomGenerators.DiscreteDistributions
 			}
 		}
 
+		/// <inheritdoc/>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public override void DropSharedGenerator()
+		{
+			m_sharedGenerator = null;
+		}
+
 		private void OnValidate()
 		{
 			m_sharedGenerator = null;