+++ Runtime/DiscreteDistributions/WeightedDistribution/GeneratorProviders/WeightedGeneratorDependentProvider.cs
@@ -1,5 +1,6 @@
 // Copyright (c) 2020 Vladimir Popov zor1994@gmail.com https://github.com/ZorPastaman/Random-Generators
 
+using System.Runtime.CompilerServices;
 using JetBrains.Annotations;
 using UnityEngine;
 using Zor.RandomGenerators.ContinuousDistributions;
@@ -79,6 +80,84 @@ namespace Zor.RandomGenerators.DiscreteDistributions
 
 				return m_sharedGenerator;
 			}
+		}
+
+		public ContinuousGeneratorProviderReference dependedGeneratorProvider
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining), Pure]
+			get => m_DependedGeneratorProvider;
+			set
+			{
+				if (m_DependedGeneratorProvider == value)
+				{
+					return;
+				}
+
+				m_DependedGeneratorProvider = value;
+				m_sharedGenerator = null;
+			}
+		}
+
+		/// <summary>
+		/// How many values are used by its generator.
+		/// </summary>
+		public int valuesCount
+		{
+			[MethodImpl(MethodImplOptions.AggressiveInlining), Pure]
+			get => m_Values.Length;
+		}
+
+		/// <summary>
+		/// Gets a value and weight at the index <paramref name="index"/>.
+		/// </summary>
+		/// <param name="index"></param>
+		/// <returns>Value and weight at the index <paramref name="index"/>.</returns>
+		[MethodImpl(MethodImplOptions.AggressiveInlining), Pure]
+		public (T, uint) GetValueWeight(int index)
+		{
+			return (m_Values[index], m_Weights[index]);
+		}
+
+		/// <summary>
+		/// Sets a value and weight at the index <paramref name="index"/>.
+		/// </summary>
+		/// <param name="value"></param>
+		/// <param name="weight"></param>
+		/// <param name="index"></param>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public void SetValueWeight([CanBeNull] T value, uint weight, int index)
+		{
+			m_Values[index] = value;
+			m_Weights[index] = weight;
+			m_sharedGenerator = null;
+		}
+
+		/// <summary>
+		/// Sets values and weights
+		/// </summary>
+		/// <param name="values"></param>
+		/// <param name="weights"></param>
+		/// <remarks>
+		/// Counts of Values and Weights must be the same.
+		/// </remarks>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public void SetValuesWeights([NotNull] T[] values, [NotNull] uint[] weights)
+		{
+			m_Values = values;
+			m_Weights = weights;
+			m_sharedGenerator = null;
+		}
+
+		/// <inheritdoc/>
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		public sealed override void DropSharedGenerator()
+		{
+			m_sharedGenerator = null;
+		}
+
+		private void OnValidate()
+		{
+			m_sharedGenerator = null;
 		}
 	}
 }